Add logic for time-related fields
Updated the popular-apps shortcode on the homepage to conditionally show an option for the newest apps whose date_added string falls within 30 days of the website build time. This option isn't shown when no app meets the criteria

Reworked the github-content a bit to show the date added and last (truenas) updated dates over in the details panel.

Finally, I pulled the date added value display from the catalog shortcode as (for now), the most recent apps display is handled on the home page and the additional detail for add date is being communicated as part of the details view.