Add featured-apps and do some code cleanup
Featured apps shortcode is for the home page, to promote up to 4 apps of our choosing.
Bring the telemetry awards behaviors in to its own partial template that can be called in all the shortcodes without any more copy/paste.
Do some code cleanup and troubleshooting on the shortcodes.
Fix a spacing bug in the footer.