PD-2166 Update the SNMD and MNMD Cluster article Certificate Info
This commit updates the two MinIO cluster tutorials certificate information with new instructions on either importing an existing certificate or creating one in TrueNAS. It adds a missing SNMD configuration snippet. It removed duplicate content about the two clusters in each of the articles and reduces it to what applies to that type of cluster.