PD-2166 Update Certificate Information
This commit removes the option to create an ACME certificate for a MNMD cluster as entering the IP address in the SAN field results in an error and failure to create the ACME DNS certificate.
Removing this from the SNMD tutorial for now and until we fully test the ACME option for clusters.
This commit changes mentions of creating a certificat to obtaining and importing one.