Pd 2166 update min io cluster tutorials (#99)
* PD-2166 Update MinIO Enterprise tutorial

This commit updates the non-cluster deployment of the MinIO Enterprise app.
It removes some of the cluster specific information, and adds new information for changes to certificates.

* PD-2166 Update the SNMD and MNMD Cluster article Certificate Info

This commit updates the two MinIO cluster tutorials certificate information with new instructions on either importing an existing certificate or creating one in TrueNAS. It adds a missing SNMD configuration snippet. It removed duplicate content about the two clusters in each of the articles and reduces it to what applies to that type of cluster.

* PD-2166 Update Certificate Information

This commit removes the option to create an ACME certificate for a MNMD cluster as entering the IP address in the SAN field results in an error and failure to create the ACME DNS certificate.
Removing this from the SNMD tutorial for now and until we fully test the ACME option for clusters.
This commit changes mentions of creating a certificat to obtaining and importing one.

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>